pyproject.toml
[project]
name = "slide-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "numpy>=2.2.1",
    "opencv-python>=4.10.0.84",
    "pillow>=11.0.0",
    "pytube>=15.0.0",
    "reportlab>=4.2.5",
    "ruff>=0.8.4",
]
